UnreadNumberObserver is now in a separated js file
